+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 
 - Added reply channel
 - impl `Default` for `Context`
+- improved cancellation ergonomics
 
 ### Breaking Changes
 
 - `context.canncelled()?` replaces `context.clone()?` to check for cancellation
+- removed `Cancelled` type
+- `Context<R>::Output: Self` - must check for cancellation at end of `try`-block, unless it contains an infinite loop
 
 ## [v0.0.1]